+++ Assets/Scripts/IDamageable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    int Health { get; set; }
+    float Health { get; set; }
     void Damage();
 
 }
